Fixing merge conflicts
