+++ src/ProcessMaker/Nayra/Bpmn/IntermediateCatchEventTrait.php
@@ -2,10 +2,7 @@
 
 namespace ProcessMaker\Nayra\Bpmn;
 
-use ProcessMaker\Nayra\Bpmn\EndTransition;
-use ProcessMaker\Nayra\Bpmn\State;
 use ProcessMaker\Nayra\Contracts\Bpmn\EventDefinitionInterface;
-use ProcessMaker\Nayra\Contracts\Bpmn\EventInterface;
 use ProcessMaker\Nayra\Contracts\Bpmn\FlowNodeInterface;
 use ProcessMaker\Nayra\Contracts\Bpmn\GatewayInterface;
 use ProcessMaker\Nayra\Contracts\Bpmn\IntermediateCatchEventInterface;
@@ -22,8 +19,7 @@ use ProcessMaker\Nayra\Contracts\Repositories\RepositoryFactoryInterface;
  */
 trait IntermediateCatchEventTrait
 {
-
-    use FlowNodeTrait;
+     use CatchEventTrait;
 
     /**
      * Receive tokens.
@@ -105,12 +101,5 @@ trait IntermediateCatchEventTrait
         $this->triggerPlace->addNewToken($instance);
         return $this;
     }
-
-    /**
-     * @return \ProcessMaker\Nayra\Engine\ExecutionInstance[]
-     */
-    public function getTargetInstances(EventDefinitionInterface $message, TokenInterface $token)
-    {
-        return $this->getOwnerProcess()->getInstances();
-    }
 }
+
