+++ src/ProcessMaker/Nayra/Contracts/Bpmn/CatchEventInterface.php
@@ -9,36 +9,6 @@ namespace ProcessMaker\Nayra\Contracts\Bpmn;
  */
 interface CatchEventInterface extends EventInterface
 {
-
-    /**
-     * Get true when all of the types of triggers that are listed in the
-     * catch Event MUST be triggered before the Process is instantiated.
-     *
-     * @return boolean
-     */
-    public function isParallelMultiple();
-
-    /**
-     * Get Data Outputs for the catch Event.
-     *
-     * @return DataOutputInterface[]
-     */
-    public function getDataOutputs();
-
-    /**
-     * Get Data Associations of the catch Event.
-     *
-     * @return DataOutputAssociationInterface[]
-     */
-    public function getDataOutputAssociations();
-
-    /**
-     * Get OutputSet for the catch Event.
-     *
-     * @return OutputSetInterface
-     */
-    public function getOutputSet();
-
     /**
      * Get EventDefinitions that are triggers expected for a catch Event.
      *
